logitech-hidpp: make receiver runtime class derivable

+++ plugins/wacom-raw/fu-wacom-device.h
@@ -6,8 +6,9 @@
 
 #pragma once
 
+#include <fwupdplugin.h>
+
 #include "fu-wacom-common.h"
-#include "fu-udev-device.h"
 
 #define FU_TYPE_WACOM_DEVICE (fu_wacom_device_get_type ())
 G_DECLARE_DERIVABLE_TYPE (FuWacomDevice, fu_wacom_device, FU, WACOM_DEVICE, FuUdevDevice)
@@ -44,5 +45,5 @@ gboolean	 fu_wacom_device_erase_all	(FuWacomDevice	*self,
 						 GError		**error);
 gboolean	 fu_wacom_device_check_mpu	(FuWacomDevice	*self,
 						 GError		**error);
-guint		 fu_wacom_device_get_block_sz	(FuWacomDevice	*self);
+gsize		 fu_wacom_device_get_block_sz	(FuWacomDevice	*self);
 guint		 fu_wacom_device_get_base_addr	(FuWacomDevice	*self);
